Use componentDidMount event to fade in view
componentWillReceiveProps is only invoked when props are changed which may not be the case
when action sheet is mounted and unmounted.
When component is hidden with opacity=0, it still receives pointer events so it will block
clicking on controls even though it's hidden. Unmounting the component is better as it won't
block